Move PoCSocket to its own target for clarity
