+++ Tests/PoCSocketTests/Helpers/SimpleResponseCreator.swift
@@ -1,0 +1,1 @@
+../../HTTPTests/Helpers/SimpleResponseCreator.swift
